requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
reportlab>=4.0.8
praw>=7.7.1
